added a basic server reload ability, using socket.io / grunt's watch
the whole idea is refering to https://github.com/tbranyen/backbone-boilerplate/issues/10

Might do a PR on backbone-boilerplate repo, but here is the the same
functionnality tailored to this grunt's build script.

A tiny client-side script is injected on every `*.html` requests, the
connect task (might be renamed) setup a basic connect server with
socket.io enabled. Whenever a new connection is established, the socket
object is stored through `config('socket')` to be able to use it and
emit from another task (a task that it triggered by grunt's watch)

kudos to @tbranyen for the original idea