usemin - reworked the usemin task to be a little bit clever
original idea from @necolas in:
https://github.com/h5bp/html5-boilerplate/issues/831

usemin task is now using "directives", some special kind of html comments
surrounding the part of html we want to replace.

There is no need for JSDOM for this, this is plain regexp (JSDOM required
for data-build attributes).

Ex:

    <!-- build:css css/site.css -->
    <link rel="stylesheet" href="css/style.css">
    <!-- endbuild -->

Results in:

    <link rel="stylesheet" href="css/site.css">

We even have tests for that :) (node test/tasks/usemin)

Relates to #13