added htmlclean task, build targets and removed tasks' subdirs
* added htmlclean task, running html through html-minifier. Closes #23

* plugin config/tasks loading now done in tasks/h5bp.js

* Added few alias as build targets. Closes #21

the grunt.js file in repo's root is there only for assisting devlopping
the project. This ensures the tasks and default configuration is loaded
globally, as a new gruntfile might override this per project. Closes #22

* pass config in a slightly different way to grunt-help plugin

we need this to remain the same, and to not rely on grunt.config('help')
as the entire config may be re-init in some other gruntfile.

* moved all tasks files into tasks/*.js (no more subdir, except the init one)