add tasks to generate test for a given task
Should this make testing easier. Added task which prompt for a bunch of
values, which are used with the gherkin feature template file in
test/features/template.feature.

A default step definition file is used (if there's no task specific
step defs) with snippets to run a given task, and some basic assertions.

The test generated should provide a good starting point but will most
likely need to adjust a few path maybe (unless you give the correct
prompt, or adjust the feature file and regenerate)

Usually, each task is comparing a file or a set of files in `.test`
directory with expected files stored in test/fixtures/<taskname>

Tests needs to define a mocha before handler with helper.before
function.